dom/workers/test/test_bug1060621.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test for URLSearchParams object in workers</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  var worker = new Worker("bug1060621_worker.js");

  worker.onmessage = function(event) {
    ok(true, "The operation is done. We should not leak.");
    SimpleTest.finish();
  };

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
